querySelector, qs, q, dom minupulator
add the q to the gobal scope, is the same that boq.utils.qs

create following the functions in dom manipulador:
html(html) set or get the html of selected nodes.
attr(name,value) set or get the html of selected nodes.
css(name,value) set or get the style rule in the selected nodes.
hide() hide with css the selected nodes.
show() show with css the selected nodes.
remove() remove from dom the selected nodes.